refactor: mass refactoring a collector logic
