+++ src/spaceone/inventory/manager/__init__.py
@@ -3,7 +3,7 @@ from spaceone.inventory.manager.identity_manager import IdentityManager
 from spaceone.inventory.manager.cloud_service_type_manager import CloudServiceTypeManager
 from spaceone.inventory.manager.cloud_service_manager import CloudServiceManager
 from spaceone.inventory.manager.collection_state_manager import CollectionStateManager
-from spaceone.inventory.manager.collector_manager.__init__ import *
+from spaceone.inventory.manager.collector_manager import *
 from spaceone.inventory.manager.cleanup_manager import CleanupManager
 from spaceone.inventory.manager.resource_group_manager import ResourceGroupManager
 from spaceone.inventory.manager.config_manager import ConfigManager
@@ -11,3 +11,5 @@ from spaceone.inventory.manager.note_manager import NoteManager
 from spaceone.inventory.manager.change_history_manager import ChangeHistoryManager
 from spaceone.inventory.manager.record_manager import RecordManager
 from spaceone.inventory.manager.collector_rule_manager import CollectorRuleManager
+from spaceone.inventory.manager.secret_manager import SecretManager
+from spaceone.inventory.manager.collector_plugin_manager import CollectorPluginManager
