+++ src/spaceone/inventory/manager/repository_manager.py
@@ -22,4 +22,3 @@ class RepositoryManager(BaseManager):
 
         if version not in response.get('results', []):
             raise ERROR_INVALID_PLUGIN_VERSION(plugin_id=plugin_id, version=version)
-
