+++ src/spaceone/inventory/manager/schedule_manager.py
@@ -1,3 +1,6 @@
+"""
+Deprecated
+"""
 import logging
 
 from spaceone.core.manager import BaseManager
@@ -7,8 +10,6 @@ _LOGGER = logging.getLogger(__name__)
 """
 Schedule
 """
-
-
 class ScheduleManager(BaseManager):
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
